Creacion de nuevas policies, ajuste en los index.blade para la utilizacion de permisos, nueva migracion que añade branch_id payments

+++ app/Http/Controllers/Web/BranchWebController.php
@@ -4,12 +4,14 @@ namespace App\Http\Controllers\Web;
 
 use App\Models\Province;
 use App\Http\Controllers\BaseBranchController;
+use App\Models\Branch;
 use Illuminate\Http\Request;
 
 class BranchWebController extends BaseBranchController
 {
     public function index()
     {
+        $this->authorize('viewAny', Branch::class);
         $branches = $this->branchService->getAllBranches();
         $rowData = $this->branchService->getAllBranchesForDatatable();
 
@@ -21,12 +23,14 @@ class BranchWebController extends BaseBranchController
 
     public function create()
     {
+        $this->authorize('create', Branch::class);
         $provinces = Province::orderBy('name')->get();
         return view('admin.branch.create', compact('provinces'));
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create', Branch::class);
         try {
             $branch = $this->branchService->createBranch($request->all());
             return redirect()->route('web.branches.index')
@@ -41,6 +45,7 @@ class BranchWebController extends BaseBranchController
     public function edit($id)
     {
         $branch = $this->branchService->getBranchById($id);
+        $this->authorize('update', $branch);
         $provinces = Province::orderBy('name')->get();
 
         return view('admin.branch.edit', compact('branch', 'provinces'));
@@ -48,8 +53,10 @@ class BranchWebController extends BaseBranchController
 
     public function update(Request $request, $id)
     {
+        $branch = $this->branchService->getBranchById($id);
+        $this->authorize('update', $branch);
         try {
-            $branch = $this->branchService->updateBranch($id, $request->all());
+            $this->branchService->updateBranch($id, $request->all());
             return redirect()->route('web.branches.index')
                 ->with('success', 'Sucursal actualizada exitosamente');
         } catch (\Illuminate\Validation\ValidationException $e) {
@@ -61,6 +68,8 @@ class BranchWebController extends BaseBranchController
 
     public function destroy($id)
     {
+        $branch = $this->branchService->getBranchById($id);
+        $this->authorize('delete', $branch);
         $this->branchService->deleteBranch($id);
         return redirect()->route('web.branches.index')
             ->with('success', 'Sucursal eliminada exitosamente');
